requirements.txt
django>=3.0.2,<3.1.0
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.8.0,<2.8.4
flake8>=3.6.0,<3.7.0